# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "网络数字孪生平台 REST API 文档。",
        "title": "Network Digital Twin API",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/device/{connector}/{query_type}": {
            "get": {
                "description": "根据连接器和查询类型查询设备信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "查询设备信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "connector name",
                        "name": "connector",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query type",
                        "name": "query_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "device name",
                        "name": "device",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/monitor/{connector}/{query_type}": {
            "get": {
                "description": "根据连接器和查询类型查询监控数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitor"
                ],
                "summary": "查询监控数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "connector name",
                        "name": "connector",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query type",
                        "name": "query_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "device name",
                        "name": "device",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "port name",
                        "name": "port",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "metrics list",
                        "name": "metrics",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end time",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                400001,
                401001,
                404001,
                429001,
                500001,
                501001,
                503001,
                400011,
                503011,
                400021,
                404021,
                400031,
                500031,
                400041,
                404041,
                501041,
                400051,
                404051,
                501051
            ],
            "x-enum-comments": {
                "CodeBadRequest": "请求参数错误",
                "CodeCircuitBreakOpen": "服务不可用（熔断）",
                "CodeDeviceBadRequest": "设备请求参数错误",
                "CodeDeviceNotFound": "Connector 未找到",
                "CodeDeviceUnsupported": "设备操作不支持",
                "CodeInternalError": "服务内部错误",
                "CodeMonitorBadRequest": "监控请求参数错误",
                "CodeMonitorNotFound": "监控 Connector 未找到",
                "CodeMonitorUnsupported": "监控操作不支持",
                "CodeNotFound": "资源未找到",
                "CodeNotImplemented": "功能未实现",
                "CodeRateLimitExceed": "请求限流",
                "CodeSnapshotBadRequest": "快照请求参数错误",
                "CodeSnapshotNotFound": "快照未找到",
                "CodeSuccess": "成功",
                "CodeSyncQueueFull": "Webhook 队列满",
                "CodeSyncUnsupportedAction": "不支持的同步动作",
                "CodeTopologyBadRequest": "拓扑请求参数错误",
                "CodeTopologyQueryFailed": "拓扑查询失败",
                "CodeUnauthorized": "未认证"
            },
            "x-enum-descriptions": [
                "成功",
                "请求参数错误",
                "未认证",
                "资源未找到",
                "请求限流",
                "服务内部错误",
                "功能未实现",
                "服务不可用（熔断）",
                "不支持的同步动作",
                "Webhook 队列满",
                "快照请求参数错误",
                "快照未找到",
                "拓扑请求参数错误",
                "拓扑查询失败",
                "设备请求参数错误",
                "Connector 未找到",
                "设备操作不支持",
                "监控请求参数错误",
                "监控 Connector 未找到",
                "监控操作不支持"
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeNotFound",
                "CodeRateLimitExceed",
                "CodeInternalError",
                "CodeNotImplemented",
                "CodeCircuitBreakOpen",
                "CodeSyncUnsupportedAction",
                "CodeSyncQueueFull",
                "CodeSnapshotBadRequest",
                "CodeSnapshotNotFound",
                "CodeTopologyBadRequest",
                "CodeTopologyQueryFailed",
                "CodeDeviceBadRequest",
                "CodeDeviceNotFound",
                "CodeDeviceUnsupported",
                "CodeMonitorBadRequest",
                "CodeMonitorNotFound",
                "CodeMonitorUnsupported"
            ]
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0 = 成功",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ErrorCode"
                        }
                    ]
                },
                "data": {},
                "message": {
                    "description": "人类可读描述",
                    "type": "string"
                }
            }
        }
    }
}